+++ lib/controller.js
@@ -17,7 +17,7 @@ const ExtensionDeviceReceive = require('./extension/deviceReceive');
 const ExtensionMarkOnlineXiaomi = require('./extension/markOnlineXiaomi');
 const ExtensionBridgeConfig = require('./extension/bridgeConfig');
 const ExtensionGroups = require('./extension/groups');
-const DeviceAvailabilityHandler = require('./extension/deviceAvailabilityHandler');
+const DeviceAvailability = require('./extension/deviceAvailability');
 
 class Controller {
     constructor() {
@@ -42,7 +42,6 @@ class Controller {
             new ExtensionMarkOnlineXiaomi(this.zigbee, this.mqtt, this.state, this.publishDeviceState),
             new ExtensionBridgeConfig(this.zigbee, this.mqtt, this.state, this.publishDeviceState),
             new ExtensionGroups(this.zigbee, this.mqtt, this.state, this.publishDeviceState),
-            new DeviceAvailabilityHandler(this.zigbee, this.mqtt, this.state, this.publishDeviceState),
         ];
 
         if (settings.get().homeassistant) {
@@ -53,6 +52,12 @@ class Controller {
 
         if (settings.get().advanced.soft_reset_timeout !== 0) {
             this.extensions.push(new ExtensionSoftReset(
+                this.zigbee, this.mqtt, this.state, this.publishDeviceState
+            ));
+        }
+
+        if (settings.get().experimental.availablility_timeout) {
+            this.extensions.push(new DeviceAvailability(
                 this.zigbee, this.mqtt, this.state, this.publishDeviceState
             ));
         }
@@ -142,13 +147,6 @@ class Controller {
                 }
             });
         });
-    }
-
-    onPublishDeviceState(device, messagePayload) {
-        // Call extensions.
-        this.extensions
-            .filter((e) => e.onPublishDeviceState)
-            .forEach((e) => e.onPublishDeviceState(device, messagePayload));
     }
 
     stop(callback) {
@@ -243,8 +241,6 @@ class Controller {
         }
 
         this.mqtt.publish(friendlyName, JSON.stringify(messagePayload), options);
-
-        this.onPublishDeviceState(device, messagePayload);
     }
 
     getDeviceInfoForMqtt(device) {
