+++ lib/extension/deviceAvailability.js
