+++ lib/extension/homeassistant.js
@@ -500,23 +500,12 @@ class HomeAssistant {
             const topic = `${config.type}/${ieeeAddr}/${config.object_id}/config`;
             const payload = {...config.discovery_payload};
             payload.state_topic = `${settings.get().mqtt.base_topic}/${friendlyName}`;
-            payload.availability_topic = `${settings.get().mqtt.base_topic}/bridge/state`;
 
             // Set (unique) name
             payload.name = `${friendlyName}_${config.object_id}`;
 
             // Set unique_id
             payload.unique_id = `${ieeeAddr}_${config.object_id}_${settings.get().mqtt.base_topic}`;
-
-            // Add last_message to attributes
-            if (settings.get().advanced.add_timestamp) {
-                if (!payload.hasOwnProperty('json_attributes')) {
-                    payload.json_attributes = [];
-                }
-                if (!payload.json_attributes.includes('last_message')) {
-                    payload.json_attributes.push('last_message');
-                }
-            }
 
             // Attributes for device registry
             payload.device = {
@@ -527,15 +516,13 @@ class HomeAssistant {
                 manufacturer: mappedModel.vendor,
             };
 
-            // Set availability payload
-            const attributeReportInterval = settings.getDevice(ieeeAddr).hasOwnProperty('attribute_report_interval');
-            if (settings.get().advanced.hasOwnProperty('availability_interval') || attributeReportInterval) {
-                const device = this.zigbee.getDevice(ieeeAddr);
-                if (attributeReportInterval || (device && (device.powerSource === 'Mains (single phase)'))) {
-                    payload.availability_topic = `${settings.get().mqtt.base_topic}/devices/${friendlyName}/state`;
-                    payload.payload_not_available = 'offline';
-                    payload.payload_available = 'online';
-                }
+            // Set availablility payload
+            // When using experimental availablility_timeout each device has it's own availablility topic.
+            // If not, use the availablility topic of zigbee2mqtt.
+            if (settings.get().experimental.availablility_timeout) {
+                payload.availability_topic = `${settings.get().mqtt.base_topic}/${friendlyName}/availablility`;
+            } else {
+                payload.availability_topic = `${settings.get().mqtt.base_topic}/bridge/state`;
             }
 
             if (payload.command_topic) {
