+++ lib/util/settings.js
@@ -39,6 +39,9 @@ const defaults = {
          */
         network_key: [1, 3, 5, 7, 9, 11, 13, 15, 0, 2, 4, 6, 8, 10, 12, 13],
     },
+    experimental: {
+        availablility_timeout: 0,
+    },
 };
 
 let settings = read();
